Close #8040: Move ImageLoader to support-base
The `ImageLoader` API is public and we should have it in a better place.
