+++ components/browser/tabstray/build.gradle
@@ -31,7 +31,7 @@ dependencies {
 
     implementation project(':ui-icons')
     implementation project(':ui-colors')
-    implementation project(':support-base')
+    implementation project(':concept-base')
     implementation project(':support-images')
     implementation project(':support-ktx')
