+++ components/browser/thumbnails/src/main/java/mozilla/components/browser/thumbnails/storage/ThumbnailStorage.kt
@@ -18,8 +18,8 @@ import mozilla.components.browser.thumbnails.R
 import mozilla.components.browser.thumbnails.utils.ThumbnailDiskCache
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.images.DesiredSize
-import mozilla.components.support.images.ImageLoadRequest
-import mozilla.components.support.images.ImageSaveRequest
+import mozilla.components.concept.base.images.ImageLoadRequest
+import mozilla.components.concept.base.images.ImageSaveRequest
 import mozilla.components.support.images.decoder.AndroidImageDecoder
 import java.util.concurrent.Executors
