+++ components/concept/base/build.gradle
@@ -27,6 +27,7 @@ android {
 dependencies {
     implementation Dependencies.kotlin_stdlib
     implementation Dependencies.kotlin_coroutines
+    implementation Dependencies.androidx_annotation
 
     testImplementation Dependencies.androidx_test_junit
     testImplementation Dependencies.testing_robolectric
