+++ components/support/images/build.gradle
@@ -37,7 +37,6 @@ android {
 dependencies {
     implementation project(':support-base')
 
-    implementation Dependencies.androidx_annotation
     implementation Dependencies.kotlin_stdlib
     implementation Dependencies.kotlin_coroutines
